requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
defusedxml==0.7.1
dj-database-url==3.0.0
Django>=4.2,<5.0
et-xmlfile==2.0.0
fonttools==4.58.4
fpdf2==2.7.4
gunicorn==22.0.0
idna==3.10
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.0.3
pillow==11.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.6.0
xlrd>=2.0.1